requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
groq==0.4.1
jsonschema==4.19.1
tiktoken==0.5.1
python-multipart==0.0.6
python-dotenv==1.0.0